--- sagas/auth/getCredentialFromFacebook.js
@@ -1,25 +1,0 @@
-import { call, put } from 'redux-saga/effects';
-import { auth } from '../../services';
-import utils from '../../utils';
-
-export default function* getCredentialFromFacebook(action) {
-  try {
-    const response = yield call(auth.getCredentialFromFacebook);
-    const nextAction = utils.app.prepareNextAction(action, response);
-
-    if (nextAction) {
-      yield put(nextAction);
-    } else if (response) {
-      yield put({
-        type: 'SIGN_IN_USER',
-        payload: response,
-      });
-    }
-  } catch (error) {
-    yield put({
-      type: 'SET_SYSTEM_MESSAGE',
-      payload: utils.app.createError(error),
-      error: true,
-    });
-  }
-}
