+++ sagas/auth/signOut/index.js
@@ -1,0 +1,25 @@
+import { call, put } from 'redux-saga/effects';
+
+import { auth } from '../../../services';
+import utils from '../../../utils';
+
+export default function* signOut(action) {
+  try {
+    const response = yield call(auth.signOut);
+    const nextAction = utils.app.prepareNextAction(action, response);
+
+    if (nextAction) {
+      yield put(nextAction);
+    } else {
+      yield put({
+        type: 'SIGN_OUT_USER',
+      });
+    }
+  } catch (error) {
+    yield put({
+      type: 'SET_SYSTEM_MESSAGE',
+      payload: utils.app.createError(error),
+      error: true,
+    });
+  }
+}
